add setup.py
